feat(power): fold pmic drivers in, add M5PM1 backend, migrate M5Stack
- Dissolve src/peripherals/pmic into the power component: the axp2101 and m5pm1
  chip drivers move under power/tdd_power/src/<chip>/ (no standalone pmic
  component). Build them with ENABLE_POWER; drop ENABLE_PMIC*.
- Add the tdd_power_m5pm1 backend (battery from the M5PM1 ADC, charge state from
  the power source, plug/unplug IRQ on GPIO13) and register the power device on
  M5STACK_STICKS3. Slim the M5Stack board file (internal static bring-up, drop
  the unused public power API and the display test helper) and fix its display
  select (ENABLE_ESP_DISPLAY -> ENABLE_DISPLAY).
- Migrate POCKET / WAVESHARE_ESP32S3 board Kconfig ENABLE_PMIC* -> ENABLE_POWER.
- power selftest: portable format specifiers (cast u32/u8 to unsigned) + add
  M5STACK configs.
- Adopt the platform flash-size derivation: bump the ESP32 platform pin and drop
  the now-redundant PLATFORM_FLASHSIZE selects on ESP32-C6 / ESP32-P4-C6.

+++ examples/peripherals/power/selftest/src/power_selftest.c
@@ -46,8 +46,11 @@ static void __screen_update(void)
     if (NULL == g_label) {
         return;
     }
-    snprintf(buf, sizeof(buf), "POWER SELF-TEST\n%u.%02uV   %u%%\n[ %s ]\n%s\nevt %d    P%d  F%d", g_mv / 1000,
-             (g_mv % 1000) / 10, g_pct, __chg_str(g_chg), g_phase, g_evt_cnt, g_pass, g_fail);
+    // Cast the u32/u8 values to unsigned: uint32_t is unsigned int on some SoCs (T5)
+    // and unsigned long on others (ESP32), so %u alone is not portable.
+    snprintf(buf, sizeof(buf), "POWER SELF-TEST\n%u.%02uV   %u%%\n[ %s ]\n%s\nevt %d    P%d  F%d",
+             (unsigned)(g_mv / 1000), (unsigned)((g_mv % 1000) / 10), (unsigned)g_pct, __chg_str(g_chg), g_phase,
+             g_evt_cnt, g_pass, g_fail);
 
     lv_vendor_disp_lock();
     lv_label_set_text(g_label, buf);
@@ -162,8 +165,9 @@ void user_main(void)
     ui_phase("battery info");
     rt = tdl_power_get_info(h, &info);
     if (OPRT_OK == rt) {
-        PR_NOTICE("  landmarks: full=%u empty=%u low=%u critical=%u mV, curve=%u", info.battery.v_full_mv,
-                  info.battery.v_empty_mv, info.battery.v_low_mv, info.battery.v_critical_mv, info.battery.curve_cnt);
+        PR_NOTICE("  landmarks: full=%u empty=%u low=%u critical=%u mV, curve=%u", (unsigned)info.battery.v_full_mv,
+                  (unsigned)info.battery.v_empty_mv, (unsigned)info.battery.v_low_mv,
+                  (unsigned)info.battery.v_critical_mv, (unsigned)info.battery.curve_cnt);
         CHECK(info.battery.v_full_mv > info.battery.v_empty_mv, "v_full > v_empty");
     }
 
@@ -176,10 +180,10 @@ void user_main(void)
         PR_NOTICE("  no battery on this board");
     } else {
         g_mv = mv;
-        CHECK(OPRT_OK == rt && mv > 2000 && mv < 5000, "voltage plausible (2-5V): %u mV", mv);
+        CHECK(OPRT_OK == rt && mv > 2000 && mv < 5000, "voltage plausible (2-5V): %u mV", (unsigned)mv);
         rt = tdl_power_battery_get_percent(h, &pct);
         g_pct = pct;
-        CHECK(OPRT_OK == rt && pct <= 100, "percent 0-100: %u%%", pct);
+        CHECK(OPRT_OK == rt && pct <= 100, "percent 0-100: %u%%", (unsigned)pct);
     }
 
     /* charger state */
@@ -232,7 +236,7 @@ void user_main(void)
     while (1) {
         __read_live(h);
         __screen_update();
-        PR_NOTICE("[live] %u mV, %u%%, %s", g_mv, g_pct, __chg_str(g_chg));
+        PR_NOTICE("[live] %u mV, %u%%, %s", (unsigned)g_mv, (unsigned)g_pct, __chg_str(g_chg));
         tal_system_sleep(5000);
     }
 }
